functions/process_json/requirements.txt
# Dependências para Cloud Functions
google-cloud-storage
